refactor/15-move-hr_addon_settings.py-to-workday.py-v15
